Create login/signup endpoints (#21)
* Implement signup/login endpoints

Users can signup and login to the system. Created two instances of Mongo connections for 'user_information' and 'fashion_images' collections. Users cannot signup with a username/email that already exists in the database.  Password of the user is encrypted using flask-bcrypt before insertion to the DB.  A session with the logged-in/registered username is returned in each endpoint upon success. Tested using local html forms (not included).

* Remove the comments for redirection