Fix signup requests
I did not come across CORS issue once I fixed my signUpUser request. 
I realized that I tried add the whole form to the fmData object, which resulted in an empty object on the server side, fixed it by appending the [key, value] pairs one by one, tried for loop and map functions but could not get it to work.