Login signup frontend (#23)
* Add antd login template

* Add dummy signup code

* Create login/signup endpoints (#21)

* Implement signup/login endpoints

Users can signup and login to the system. Created two instances of Mongo connections for 'user_information' and 'fashion_images' collections. Users cannot signup with a username/email that already exists in the database.  Password of the user is encrypted using flask-bcrypt before insertion to the DB.  A session with the logged-in/registered username is returned in each endpoint upon success. Tested using local html forms (not included).

* Remove the comments for redirection

* Add name and surname fields in the signup backend

* Add login and signup requests 

++Fix input fields in the signup form, add password matching rule

* Implement signup and login requests (WIP.)

* Fix signup requests

I did not come across CORS issue once I fixed my signUpUser request. 
I realized that I tried add the whole form to the fmData object, which resulted in an empty object on the server side, fixed it by appending the [key, value] pairs one by one, tried for loop and map functions but could not get it to work.

* Add access and refresh tokens for the backend

* Adjust formData for Login

Co-authored-by: Ata Atasoy <40474912+AtaAtasoy@users.noreply.github.com>
Co-authored-by: Ata Atasoy <h.ataatasoy@gmail.com>