Implement authorization token for logged in users (#35)
* Implement flask_jwt_extended for access tokens

Initialize the access and refresh tokens and add logout, refresh endpoints. Change /getSession to /getCurrentUser. /getCurrentUser returns the currently logged in user.

* Change getSession() to getCurrentUser()

sends a request to the /getCurrentSession endpoint with the Authentication header.

* Remove unused imports and hooks, add useEffect

UseEffect is used to display the currently logged in user.

* Implement login request and redirect to homepage

Also removed the unused imports.

* Create UserContext to store the logged in user.

* Removed the expiration time on access_token

* Add [location] on useEffect hook.

* Fixed warnings

* Supress setIsShown not used

the isShown value is not being updated anywhere, thus i supressed this warning by placing its constant value to the props of the link element.